rewrite: include vfs_mode in workdb key; always check files
We need to track the the vfs_mode in order to distinguish a
pipe/device/etc. from an empty file since they'll all have the same
oid.

Insert a row into the db for every item so that we'll know we've
already checked any whose oid doesn't change, i.e. so we don't attempt
to split the same oid more than once.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>