Fold rewrite into bup-get --pick/--force-pick/--append
Since it would be surprising for bup get to ignore the destination
repository settings (bup.split.trees, etc.) when creating new saves,
add a new --rewrite option and refuse to run when the source and
destination configs differ materially and --rewrite hasn't been
specified. Support --no-rewrite to opt out. Only the append and pick
related methods allow rewriting.

Specifying --rewrite always causes rewriting, even if the source and
destination repository settings are the same. This allows rewriting
within the same repository, and rewriting with respect to the current
defaults for things that have changed but don't have configuration
options like splitting .bupm files.

To handle the rewriting, rely on the new rewrite code, moved to
bup.rewrite, to handle the work, and remove the rewrite command since
it's equivalent to a "bup get append" to a new branch.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>