Move bup.split.trees into hashsplitter config
The hashsplitter config is now expected to include anything that could
cause a rewrite to "meaningfully" differ from the original. This is
important because we're going to rely on comparing configs in `bup
get` in order to determine what kind of transfer would be the most
appropriate default for each request.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>